requirements.txt
langchain 
streamlit 
chromadb 
PyMuPDF 
pdfplumber
groq 
python-dotenv
langchain-community 
langchain-text-splitters 
langchain-chroma 
langchain-openai
sentence-transformers
langchain-groq
